rename: Move Sidebar.jsx
